Move sorted spikes group to spikesorting
